MedSecure/requirements.txt
PyQt5>=5.15.9
cryptography>=41.0.0
Pillow>=10.0.0
numpy>=1.24.0
pyinstaller>=6.0.0
